BackEnd/Treinamento_IA/requirements.txt
xgboost==2.1.0
scikit-learn==1.5.0
pandas==2.2.0
pyarrow==17.0.0
joblib==1.4.0
mlflow==2.17.0
supabase==2.13.0
python-dotenv==1.0.0
numpy==1.26.4
